requirements.txt
pyusb==1.1.1